code-font-and-component-update
Adding a new (helper) component that receives a string (e.g. a description of a project) and wraps keywords (like 'Python') in <code> blocks so that they show up like code in the UI. I updated all component to utilize this feature (as well as other minor component changes related to the new style, the new data, and a new welcome section that includes my picture).